Module Parser.Lexer

Lexical analyser

type t

val create : unit -> t

type _ token =

| Await : [> `other ] token

| Lparen : [> `other ] token

| Rparen : [> `other ] token

| Atom : int -> [> `atom ] token

val feed : t -> char -> [ `other | `atom ] token

Feed a character to the parser.

raises Parse_error

val feed_eoi : t -> unit

Feed the end of input to the parser.

You should call this function when the end of input has been reached in order to ensure that the lexer is not awaiting more input, which would be an error.

raises Parse_error
if the lexer is awaiting more input